:art: Modify conformal files location
