+++ torch_uncertainty/post_processing/__init__.py
@@ -1,8 +1,10 @@
 # ruff: noqa: F401
 from .abstract import PostProcessing
 from .calibration import MatrixScaler, TemperatureScaler, VectorScaler
-from .conformal_aps import ConformalclassificationAPS
-from .conformal_raps import ConformalClassificationRAPS
-from .conformal_thr import ConformalclassificationTHR
+from .conformal import (
+    ConformalclassificationAPS,
+    ConformalClassificationRAPS,
+    ConformalclassificationTHR,
+)
 from .laplace import LaplaceApprox
 from .mc_batch_norm import MCBatchNorm
